support/download/git: shallow fetch of all branches
When a branch of a package is tracked using sha1 and the remote branch
moves (usually gets another commit), this script stops using a shallow
fetch since the sha1 is not anymore the head of the branch, falling back
to a full clone.

As a middle ground between a --depth 1 fetch and a full clone, fetch all
branches with depth of 100. We have a good change to catch the desired
version reducing the data transfer from the remote when fetching from
large repos, especially hardware-specific linux trees.

In the case the desired version is not in this shallow fetch, fall back
to a full fetch that reuses the objects already downloaded.

This method causes an extra burden in the server side ('Compressing
objects'), usually a few seconds, but the potential reduction of data
transferred should be beneficial for both the user and the server.

https://github.com/raspberrypi/linux.git
100 commits from all branches 567.47 MiB
git clone 1.49 GiB

https://github.com/Freescale/linux-fslc.git
100 commits from all branches 944.72 MiB + 73.89 KiB
git clone 1.68 GiB

https://github.com/linux-sunxi/sunxi-mali-proprietary
100 commits from all branches 7.55 MiB
git clone 7.55 MiB

Reported-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Ricardo Martincoski <ricardo.martincoski@datacom.ind.br>