support/download/git: optimized download of sha1
Shallow fetch don't work for downloading SHAs. However it's common to
track a tag or branch by the commit SHA so that a branch or tag doesn't
unexpectedly change.

We can take advantage of this scenario by searching the remote for a ref
that it equivalent to our sha, then shallow fetching that ref instead.

Regarding the case the remote changes between the git ls-remote and the
git fetch, once the script creates the equivalent of a shallow clone,
the fetch and the checkout do what is expected:
- for a sha1 of a removed ref (branch, tag or special ref), the fetch
  fails, falling back to the equivalent of a full clone;
- for a sha1 of a changed ref (branch, tag or special ref), the checkout
  fails (because the desired sha1 is not fetched), falling back to the
  equivalent of a full clone;
In both cases, the checkout after the equivalent of a full clone or
mirror clone is successful if the sha1 is still reachable from any ref.

Similar patch was submitted in [1] using git clone.
[1] http://patchwork.ozlabs.org/patch/681841/

Reported-by: Brandon Maier <brandon.maier@rockwellcollins.com>
Reported-by: Bryce Ferguson <bryce.ferguson@rockwellcollins.com>
Signed-off-by: Ricardo Martincoski <ricardo.martincoski@datacom.ind.br>